Unify test data for to-many
